Updated destination device default to iPhone 11 (#9)
Adds support for the default Bitrise CI device.

When Xcode 14.3 and Xcode 15 (beta) is present on the same machine, requiring the latest (this is the default) version of the Simulator runtime, will find iOS 17 (beta). This is undesirable as the beta Simulator can be unstable or at least not officially supported by earlier Xcodes.

In practice Xcode still works with newer runtimes, at least across the same major version (despite this table: https://developer.apple.com/support/xcode/). When there are different major versions, there is some risk of not working right.